Fix RecursionError in concat of LazySeq
